// src/types/public/Blogs.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ImagesId } from './Images';
import { type UsersId } from './Users';

/** Identifier type for public.blogs */
export type BlogsId = number & { __brand: 'BlogsId' };
export type BlogsName = { en: string, tr: string };

/** Represents the table public.blogs */
export default interface Blogs {
  id: BlogsId;

  title: string;
  domain: any

  image: {
    id: number,
    src: string | null,
    alt: string | null
  };

  edited_by: UsersId | null;

  table_of_contents: unknown;

  description: unknown | null;
  content: string | null;

  category: {
    id: number,
    title: string
  } | null;

  created_at: Date;

  updated_at: Date;

  cover_images: string | null
}

/** Represents the initializer for the table public.blogs */
export interface BlogsInitializer {
  /** Default value: nextval('blogs_id_seq'::regclass) */
  id?: BlogsId;

  name: unknown;

  image?: ImagesId | null;

  edited_by?: UsersId | null;

  table_of_contents: unknown;

  description?: unknown | null;

  /** Default value: '[]'::jsonb */
  relevant_branches?: unknown | null;

  /** Default value: CURRENT_TIMESTAMP */
  created_at?: Date;

  /** Default value: CURRENT_TIMESTAMP */
  updated_at?: Date;
}

/** Represents the mutator for the table public.blogs */
export interface BlogsMutator {
  id?: BlogsId;

  name?: unknown;

  image?: ImagesId | null;

  edited_by?: UsersId | null;

  table_of_contents?: unknown;

  description?: unknown | null;

  relevant_branches?: unknown | null;

  created_at?: Date;

  updated_at?: Date;
}